Remove legacy C API [AP-1684] (#1404)
# Removal Warning

This PR will be merged on **8/4/2024**. On this date the legacy C API
will become unavailable, it will not be reintroduced. On or shortly
after this date libsbp version v6.0.0 will be tagged and released

# Description

@swift-nav/devinfra



Remove legacy C API including all generated code, examples, templates,
and sbpg module.

Remove all references to "v4" API which were lying around in a handful
of places. There is no such thing as the V4 API any more, there is just
the libsbp API

# API compatibility

Yes

Legacy C API completely removed including:
- Any message types in the form `struct msg_xxx_t`
- Functions relating to sending or processing encoded payloads or frames
- payload and frame callback types
- `#define`s message types - in the form `SBP_MSG_XXX`
- C++ `LegacyState` wrapper
- All files under the `libsbp/legacy` directory
- V4 compatibility headers for modern API in `libsbp/v4/<package.h>`


## API compatibility plan

The modern API was introduce to this repository in Sep 2021. The legacy
API was retained but all symbols were marked as deprecated and compile
time messages were output by GCC and Clang whenever a legacy header was
used

In Dec 2023 the legacy was moved to a more difficult to access place. It
was no longer accessible simply by include `libsbp/sbp.h` All headers
were retained but even more compile time messages were added indicating
that these files would be deleted in the next major version

This PR is the final stage of the deprecation plan. Any users of this
repository should have already converted their code to use the moden
API.

# JIRA Reference

https://swift-nav.atlassian.net/browse/AP-1684

+++ c/test/auto_check_sbp_flash_MsgFlashErase.c
@@ -28,8 +28,8 @@ static struct {
   void *context;
 } last_msg;
 
-static u32 dummy_wr = 0;
-static u32 dummy_rd = 0;
+static size_t dummy_wr = 0;
+static size_t dummy_rd = 0;
 static u8 dummy_buff[1024];
 static void *last_io_context;
 
@@ -43,7 +43,7 @@ static void dummy_reset() {
 
 static s32 dummy_write(u8 *buff, u32 n, void *context) {
   last_io_context = context;
-  u32 real_n = n;  //(dummy_n > n) ? n : dummy_n;
+  size_t real_n = n;  //(dummy_n > n) ? n : dummy_n;
   memcpy(dummy_buff + dummy_wr, buff, real_n);
   dummy_wr += real_n;
   return (s32)real_n;
@@ -51,7 +51,7 @@ static s32 dummy_write(u8 *buff, u32 n, void *context) {
 
 static s32 dummy_read(u8 *buff, u32 n, void *context) {
   last_io_context = context;
-  u32 real_n = n;  //(dummy_n > n) ? n : dummy_n;
+  size_t real_n = n;  //(dummy_n > n) ? n : dummy_n;
   memcpy(buff, dummy_buff + dummy_rd, real_n);
   dummy_rd += real_n;
   return (s32)real_n;
@@ -132,13 +132,13 @@ START_TEST(test_auto_check_sbp_flash_MsgFlashErase) {
 
     ck_assert_msg(last_msg.msg.flash_erase.sector_num == 2222371310,
                   "incorrect value for last_msg.msg.flash_erase.sector_num, "
-                  "expected 2222371310, is %d",
-                  last_msg.msg.flash_erase.sector_num);
+                  "expected 2222371310, is %" PRId64,
+                  (int64_t)last_msg.msg.flash_erase.sector_num);
 
     ck_assert_msg(last_msg.msg.flash_erase.target == 74,
                   "incorrect value for last_msg.msg.flash_erase.target, "
-                  "expected 74, is %d",
-                  last_msg.msg.flash_erase.target);
+                  "expected 74, is %" PRId64,
+                  (int64_t)last_msg.msg.flash_erase.target);
   }
 }
 END_TEST
