+++ c/test/auto_check_sbp_observation_MsgSvConfigurationGpsDep.c
@@ -28,8 +28,8 @@ static struct {
   void *context;
 } last_msg;
 
-static u32 dummy_wr = 0;
-static u32 dummy_rd = 0;
+static size_t dummy_wr = 0;
+static size_t dummy_rd = 0;
 static u8 dummy_buff[1024];
 static void *last_io_context;
 
@@ -43,7 +43,7 @@ static void dummy_reset() {
 
 static s32 dummy_write(u8 *buff, u32 n, void *context) {
   last_io_context = context;
-  u32 real_n = n;  //(dummy_n > n) ? n : dummy_n;
+  size_t real_n = n;  //(dummy_n > n) ? n : dummy_n;
   memcpy(dummy_buff + dummy_wr, buff, real_n);
   dummy_wr += real_n;
   return (s32)real_n;
@@ -51,7 +51,7 @@ static s32 dummy_write(u8 *buff, u32 n, void *context) {
 
 static s32 dummy_read(u8 *buff, u32 n, void *context) {
   last_io_context = context;
-  u32 real_n = n;  //(dummy_n > n) ? n : dummy_n;
+  size_t real_n = n;  //(dummy_n > n) ? n : dummy_n;
   memcpy(buff, dummy_buff + dummy_rd, real_n);
   dummy_rd += real_n;
   return (s32)real_n;
@@ -135,20 +135,20 @@ START_TEST(test_auto_check_sbp_observation_MsgSvConfigurationGpsDep) {
     ck_assert_msg(
         last_msg.msg.sv_configuration_gps_dep.l2c_mask == 2808462402,
         "incorrect value for last_msg.msg.sv_configuration_gps_dep.l2c_mask, "
-        "expected 2808462402, is %d",
-        last_msg.msg.sv_configuration_gps_dep.l2c_mask);
+        "expected 2808462402, is %" PRId64,
+        (int64_t)last_msg.msg.sv_configuration_gps_dep.l2c_mask);
 
     ck_assert_msg(
         last_msg.msg.sv_configuration_gps_dep.t_nmct.tow == 0,
         "incorrect value for last_msg.msg.sv_configuration_gps_dep.t_nmct.tow, "
-        "expected 0, is %d",
-        last_msg.msg.sv_configuration_gps_dep.t_nmct.tow);
+        "expected 0, is %" PRId64,
+        (int64_t)last_msg.msg.sv_configuration_gps_dep.t_nmct.tow);
 
     ck_assert_msg(
         last_msg.msg.sv_configuration_gps_dep.t_nmct.wn == 0,
         "incorrect value for last_msg.msg.sv_configuration_gps_dep.t_nmct.wn, "
-        "expected 0, is %d",
-        last_msg.msg.sv_configuration_gps_dep.t_nmct.wn);
+        "expected 0, is %" PRId64,
+        (int64_t)last_msg.msg.sv_configuration_gps_dep.t_nmct.wn);
   }
 }
 END_TEST
